+++ GentleTouch/Interop/DualSense/AdaptiveTrigger.cs
@@ -3,9 +3,9 @@
 // Source: https://github.com/BadMagic100/DualSenseAPI/blob/master/DualSenseAPI/AdaptiveTrigger.cs
 public enum AdaptiveTriggerEffectType : byte
 {
-    ContinuousResistance = 0x01
+    Default              = 0x00
+  , ContinuousResistance = 0x01
   , SectionResistance    = 0x02
   , Vibrate              = 0x26
   , Calibrate            = 0xFC
-  , Default              = 0x00
 }